+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/common/request/SearchCriteriaRequestParam.java
@@ -21,6 +21,7 @@ package org.eclipse.tractusx.traceability.common.request;
 
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.eclipse.tractusx.traceability.common.model.BaseRequestFieldMapper;
@@ -39,7 +40,8 @@ import static java.util.Objects.isNull;
 @Data
 @AllArgsConstructor
 public class SearchCriteriaRequestParam {
-    @ArraySchema(arraySchema = @Schema(description = "Filter Criteria", additionalProperties = Schema.AdditionalPropertiesValue.FALSE, example = "owner,EQUAL,OWN"), maxItems = Integer.MAX_VALUE)
+    @Size(max = 50, message = "Do not use more than 50 filter query parameters")
+    @ArraySchema(arraySchema = @Schema(description = "Filter Criteria", additionalProperties = Schema.AdditionalPropertiesValue.FALSE, example = "owner,EQUAL,OWN"), maxItems = 50)
     private List<String> filter;
 
     public SearchCriteria toSearchCriteria(BaseRequestFieldMapper fieldMapper) {